+++ src/Exceptions/CTConfigException.php
@@ -8,6 +8,10 @@ use CTApi\CTLog;
 use RuntimeException;
 use Throwable;
 
+/**
+ * Class CTConfigException indicates a error in the CTConfig-Settings.
+ * @package CTApi\Exceptions
+ */
 class CTConfigException extends RuntimeException
 {
     public function __construct($message = "", $code = 0, Throwable $previous = null)