+++ src/Exceptions/CTConnectException.php
@@ -5,10 +5,13 @@ namespace CTApi\Exceptions;
 
 
 use CTApi\CTLog;
-use RuntimeException;
 use Throwable;
 
-class CTConnectException extends RuntimeException
+/**
+ * Class CTConnectException is a CRRequestException. Its thrown when ChurchTools server is not available.
+ * @package CTApi\Exceptions
+ */
+class CTConnectException extends CTRequestException
 {
     public function __construct($message = "", $code = 0, Throwable $previous = null)
     {