+++ tests/integration/Models/MetaTest.php
@@ -2,6 +2,8 @@
 
 namespace Tests\Integration\Models;
 
+use CTApi\Exceptions\CTAuthException;
+use CTApi\Exceptions\CTPermissionException;
 use CTApi\Models\Meta;
 use CTApi\Models\Person;
 use CTApi\Requests\EventAgendaRequest;
@@ -65,14 +67,22 @@ class MetaTest extends TestCaseAuthenticated
         $this->assertInstanceOf(Meta::class, $model->getMeta());
 
         //validate requestModifiedPerson and requestCreatedPerson
-        $creator = $model->getMeta()->requestCreatedPerson();
-        if (!is_null($creator)) {
-            $this->assertInstanceOf(Person::class, $creator);
+        try{
+            $creator = $model->getMeta()->requestCreatedPerson();
+            if (!is_null($creator)) {
+                $this->assertInstanceOf(Person::class, $creator);
+            }
+        }catch (CTPermissionException $exception){
+            // ignore
         }
 
-        $editor = $model->getMeta()->requestModifiedPerson();
-        if (!is_null($editor)) {
-            $this->assertInstanceOf(Person::class, $editor);
+        try{
+            $editor = $model->getMeta()->requestModifiedPerson();
+            if (!is_null($editor)) {
+                $this->assertInstanceOf(Person::class, $editor);
+            }
+        }catch (CTPermissionException $exception){
+            // ignore
         }
     }
 }